+++ build.gradle
@@ -1,34 +1,80 @@
-plugins {
-	id 'org.springframework.boot' version '2.7.3'
-	id 'io.spring.dependency-management' version '1.0.13.RELEASE'
-	id 'java'
-}
-
-group = 'dev.t1dmlgus'
-version = '0.0.1-SNAPSHOT'
-sourceCompatibility = '11'
-
-configurations {
-	compileOnly {
-		extendsFrom annotationProcessor
+buildscript {
+	ext {
+		springBootVersion = '2.7.2'
+	}
+	repositories {
+		mavenCentral()
+	}
+	dependencies {
+		classpath "org.springframework.boot:spring-boot-gradle-plugin:${springBootVersion}"
+		classpath "io.spring.gradle:dependency-management-plugin:1.0.12.RELEASE"
 	}
 }
 
-repositories {
-	mavenCentral()
+
+// 하위 모든 프로젝트 공통 세팅
+subprojects {
+
+	group 'dev.t1dmlgus'
+	version '0.0.1-SNAPSHOT'
+
+	apply plugin: 'java-library'
+	apply plugin: 'idea'
+	apply plugin: 'org.springframework.boot'
+	apply plugin: 'io.spring.dependency-management'
+
+
+	sourceCompatibility = '11'
+	targetCompatibility = '11'
+	compileJava.options.encoding = 'UTF-8'
+
+	repositories {
+		mavenCentral()
+	}
+
+//	configurations {
+//	compileOnly {
+//		extendsFrom annotationProcessor
+//	}
+
+	dependencies {
+
+		compileOnly 'org.projectlombok:lombok'
+		developmentOnly 'org.springframework.boot:spring-boot-devtools'
+		annotationProcessor 'org.projectlombok:lombok'
+		annotationProcessor "org.springframework.boot:spring-boot-configuration-processor"
+		testImplementation 'org.springframework.boot:spring-boot-starter-test'
+
+	}
+
+	test {
+		useJUnitPlatform()
+	}
 }
 
-dependencies {
-	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
-	implementation 'org.springframework.boot:spring-boot-starter-web'
-	compileOnly 'org.projectlombok:lombok'
-	developmentOnly 'org.springframework.boot:spring-boot-devtools'
-	runtimeOnly 'com.h2database:h2'
-	runtimeOnly 'mysql:mysql-connector-java'
-	annotationProcessor 'org.projectlombok:lombok'
-	testImplementation 'org.springframework.boot:spring-boot-starter-test'
+
+project(':order-api-server') {
+
+	dependencies {
+		api project(':order-core')
+	}
 }
 
-tasks.named('test') {
-	useJUnitPlatform()
+project(':order-core') {
+
+	dependencies {
+		api project(':order-common')
+	}
 }
+
+/**
+ *
+ * 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+ implementation 'org.springframework.boot:spring-boot-starter-web'
+ compileOnly 'org.projectlombok:lombok'
+ developmentOnly 'org.springframework.boot:spring-boot-devtools'
+ runtimeOnly 'com.h2database:h2'
+ runtimeOnly 'mysql:mysql-connector-java'
+ annotationProcessor 'org.projectlombok:lombok'
+ testImplementation 'org.springframework.boot:spring-boot-starter-test'
+ */
