+++ order-api-server/build.gradle
@@ -1,0 +1,20 @@
+plugins {
+    id 'java'
+}
+
+group 'dev.t1dmlgus'
+version '0.0.1-SNAPSHOT'
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+    implementation 'org.springframework.boot:spring-boot-starter-validation'
+}
+
+test {
+    useJUnitPlatform()
+}
