+++ order-api-server/src/main/java/dev/t1dmlgus/OrderMvp2Application.java
@@ -1,0 +1,13 @@
+package dev.t1dmlgus;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+
+@SpringBootApplication
+public class OrderMvp2Application {
+
+    public static void main(String[] args) {
+        SpringApplication.run(OrderMvp2Application.class, args);
+    }
+}
