+++ order-api-server/src/main/resources/application.yml
@@ -1,0 +1,21 @@
+spring:
+  datasource:
+    url: jdbc:h2:mem:order-mvp2
+    username: sa
+    password:
+    driver-class-name: org.h2.Driver
+
+#  datasource:
+#    url: jdbc:mysql://localhost:3306/order-mvp2?serverTimezone=Asia/Seoul&characterEncoding=UTF-8
+#    username: root
+#    password: 1234
+#    driver-class-name: com.mysql.cj.jdbc.Driver
+
+  jpa:
+    hibernate:
+      ddl-auto: create
+    properties:
+      hibernate:
+        show_sql: true
+        format_sql: true
+        use_sql_comments: true
