+++ order-common/build.gradle
@@ -1,0 +1,18 @@
+plugins {
+    id 'java'
+}
+
+group 'dev.t1dmlgus'
+version '0.0.1-SNAPSHOT'
+
+repositories {
+    mavenCentral()
+}
+
+dependencies {
+
+}
+
+test {
+    useJUnitPlatform()
+}
