+++ order-core/build.gradle
@@ -1,0 +1,13 @@
+
+dependencies {
+    
+    implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
+
+    runtimeOnly 'com.h2database:h2'
+    runtimeOnly 'mysql:mysql-connector-java'
+
+
+}
+
+bootJar { enabled = false }
+jar { enabled = true }
