+++ order-core/src/main/java/dev/t1dmlgus/common/MoneyConverter.java
@@ -1,0 +1,17 @@
+package dev.t1dmlgus.common;
+
+import javax.persistence.AttributeConverter;
+
+public class MoneyConverter implements AttributeConverter<Money, Integer> {
+
+
+    @Override
+    public Integer convertToDatabaseColumn(Money money) {
+        return money == null ? null : money.getValue();
+    }
+
+    @Override
+    public Money convertToEntityAttribute(Integer value) {
+        return value == null ? null : new Money(value);
+    }
+}
