+++ order-core/src/main/java/dev/t1dmlgus/config/JpaConfig.java
@@ -1,0 +1,11 @@
+package dev.t1dmlgus.config;
+
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+
+@EnableJpaAuditing
+@Configuration
+public class JpaConfig {
+}
