+++ order-core/src/main/java/dev/t1dmlgus/member/domain/Member.java
@@ -1,0 +1,28 @@
+package dev.t1dmlgus.member.domain;
+
+
+import dev.t1dmlgus.common.AbstractEntity;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@NoArgsConstructor
+@Entity
+@Table(name = "members")
+public class Member extends AbstractEntity {
+
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String memberToken;
+
+    private String name;
+    private String email;
+
+    public Member(String name, String email) {
+        this.name = name;
+        this.email = email;
+    }
+}
