+++ order-core/src/main/java/dev/t1dmlgus/member/domain/MemberRepository.java
@@ -1,0 +1,6 @@
+package dev.t1dmlgus.member.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface MemberRepository extends JpaRepository<Member, Long> {
+}
