+++ order-core/src/main/java/dev/t1dmlgus/order/domain/OrderRepository.java
@@ -1,0 +1,10 @@
+package dev.t1dmlgus.order.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface OrderRepository extends JpaRepository<Order, Long> {
+
+    Optional<Order> findByOrderToken(String orderToken);
+}
