+++ order-core/src/main/java/dev/t1dmlgus/product/domain/ProductRepository.java
@@ -1,0 +1,10 @@
+package dev.t1dmlgus.product.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface ProductRepository extends JpaRepository<Product, Long> {
+
+    Optional<Product> findByProductToken(String productToken);
+}
