+++ settings.gradle
@@ -1,1 +1,5 @@
 rootProject.name = 'order-mvp2'
+include 'order-api-server'
+include 'order-core'
+include 'order-common'
+
