--- src/main/java/dev/t1dmlgus/ordermvp2/OrderMvp2Application.java
@@ -1,13 +1,0 @@
-package dev.t1dmlgus.ordermvp2;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class OrderMvp2Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(OrderMvp2Application.class, args);
-	}
-
-}
