--- src/main/java/dev/t1dmlgus/ordermvp2/config/JpaConfig.java
@@ -1,11 +1,0 @@
-package dev.t1dmlgus.ordermvp2.config;
-
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-
-@EnableJpaAuditing
-@Configuration
-public class JpaConfig {
-}
