--- src/main/java/dev/t1dmlgus/ordermvp2/member/domain/MemberRepository.java
@@ -1,6 +1,0 @@
-package dev.t1dmlgus.ordermvp2.member.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MemberRepository extends JpaRepository<Member, Long> {
-}
