--- src/main/java/dev/t1dmlgus/ordermvp2/product/domain/ProductRepository.java
@@ -1,10 +1,0 @@
-package dev.t1dmlgus.ordermvp2.product.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface ProductRepository extends JpaRepository<Product, Long> {
-
-    Optional<Product> findByProductToken(String productToken);
-}
